+++ packages/react-app-super-template/src/SetDocumentLang.tsx
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+import { useIntl } from 'react-intl';
+
+export default function SetDocumentLang() {
+  const { locale } = useIntl();
+
+  useEffect(() => {
+    document.documentElement.lang = locale.split('-')[0]!;
+  }, [locale]);
+
+  return null;
+}